TheGuardian/requirements.txt
ultralytics
gradio
opencv-python
pillow
numpy
torch
torchvision
